Correcting block-comment-decorator setting for c-mode.
Changing the erroneous setting of the block-comment-decorator string
for C-like modes. Previously it was set on changing to a
C-mode. However, if multiple buffers with different modes were open,
this would set the variable for all of them. It has been corrected
such that the hook is added on "before-change-hook." The associated
function checks if buffer is in C-mode, and changes the variable
appropriately. Otherwise, it ensures that the variable is set to an
otherwise appropriate value.